rename step form components accordingly
- `FormRenameStep` is now named `RenameStepForm`
- `FormFillnaStep` is not named `FillnaStepForm`

Change filenames accordingly.
