+++ .storybook/bundle.ts
@@ -5,10 +5,10 @@
 
 import DataViewer from '../src/components/DataViewer.vue';
 import DomainSelector from '../src/components/DomainSelector.vue';
-import FormRenameStep from '../src/components/FormRenameStep.vue';
+import RenameStepForm from '../src/components/RenameStepForm.vue';
 import Pipeline from '../src/components/Pipeline.vue';
 import ResizablePanels from '../src/components/ResizablePanels.vue';
 import Step from '../src/components/Step.vue';
 
-export { DataViewer, DomainSelector, FormRenameStep, Pipeline, ResizablePanels, Step };
+export { DataViewer, DomainSelector, RenameStepForm, Pipeline, ResizablePanels, Step };
 export { setupStore } from '../src/store';
