+++ src/components/FillnaStepForm.vue
@@ -1,22 +1,17 @@
 <template>
   <div>
     <div class="step-edit-form">
-      <h1 class="step-edit-form__title">EDIT RENAME STEP</h1>
+      <h1 class="step-edit-form__title">FILL NULL VALUES STEP</h1>
     </div>
     <WidgetAutocomplete
-      id="oldnameInput"
-      v-model="step.oldname"
-      name="Old name"
+      id="columnInput"
+      v-model="step.column"
+      name="Fill null values in:"
       :options="columnNames"
-      @input="setSelectedColumns({ column: step.oldname })"
-      placeholder="Enter the old column name"
+      @input="setSelectedColumns({ column: step.column })"
+      placeholder="Enter a column"
     ></WidgetAutocomplete>
-    <WidgetInputText
-      id="newnameInput"
-      v-model="step.newname"
-      name="New name"
-      placeholder="Enter a new column name"
-    ></WidgetInputText>
+    <WidgetInputText id="valueInput" v-model="step.value" name="With:" placeholder="Enter a value"></WidgetInputText>
     <div class="widget-form-action">
       <button
         class="widget-form-action__button widget-form-action__button--validate"
@@ -38,36 +33,36 @@
 <script lang="ts">
 import _ from 'lodash';
 import { Mixins, Prop, Watch } from 'vue-property-decorator';
-import FormMixin, { VqbError } from '@/mixins/FormMixin.vue';
+import FormMixin from '@/mixins/FormMixin.vue';
 import { Pipeline } from '@/lib/steps';
-import renameSchema from '@/assets/schemas/rename-step__schema.json';
+import fillnaSchema from '@/assets/schemas/fillna-step__schema.json';
 import WidgetInputText from './WidgetInputText.vue';
 import WidgetAutocomplete from '@/components/WidgetAutocomplete.vue';
 import { Getter, Mutation, State } from 'vuex-class';
 import { StepFormComponent } from '@/components/formlib';
 
-interface RenameStepConf {
-  oldname: string;
-  newname: string;
+interface FillnaStepConf {
+  column: string;
+  value: string;
 }
 
 @StepFormComponent({
-  vqbstep: 'rename',
-  name: 'form-rename-step',
+  vqbstep: 'fillna',
+  name: 'fillna-step-form',
   components: {
     WidgetAutocomplete,
     WidgetInputText,
   },
 })
-export default class RenameStepForm extends Mixins(FormMixin) {
+export default class FillnaStepForm extends Mixins(FormMixin) {
   @Prop({
     type: Object,
     default: () => ({
-      oldname: '',
-      newname: '',
+      column: '',
+      value: '',
     }),
   })
-  initialValue!: RenameStepConf;
+  initialValue!: FillnaStepConf;
 
   @Prop({
     type: Boolean,
@@ -75,7 +70,7 @@ export default class RenameStepForm extends Mixins(FormMixin) {
   })
   isStepCreation!: boolean;
 
-  step: RenameStepConf = { ...this.initialValue };
+  step: FillnaStepConf = { ...this.initialValue };
 
   @State pipeline!: Pipeline;
   @State selectedStepIndex!: number;
@@ -90,30 +85,22 @@ export default class RenameStepForm extends Mixins(FormMixin) {
   @Watch('selectedColumns')
   onSelectedColumnsChanged(val: string[], oldVal: string[]) {
     if (!_.isEqual(val, oldVal)) {
-      this.step.oldname = val[0];
+      this.step.column = val[0];
     }
   }
 
   created() {
-    this.schema = renameSchema;
-    this.setSelectedColumns({ column: this.initialValue.oldname });
+    this.schema = fillnaSchema;
+    this.setSelectedColumns({ column: this.initialValue.column });
   }
 
   validateStep() {
     const ret = this.validator(this.step);
     if (ret === false) {
       this.errors = this.validator.errors;
-    } else if (this.columnNames.includes(this.step.newname)) {
-      const err: VqbError = {
-        keyword: 'nameAlreadyUsed',
-        dataPath: '.newname',
-        message: 'This column name is already used.',
-      };
-      this.errors = [err];
     } else {
       this.errors = null;
-      this.$emit('formSaved', { name: 'rename', ...this.step });
-      this.setSelectedColumns({ column: this.step.newname });
+      this.$emit('formSaved', { name: 'fillna', ...this.step });
     }
   }
