+++ src/main.ts
@@ -9,8 +9,7 @@ export { setupStore } from '@/store';
 
 // export Vue components
 import DataViewer from './components/DataViewer.vue';
-import FormRenameStep from './components/FormRenameStep.vue';
 import Pipeline from './components/Pipeline.vue';
 import ResizablePanels from './components/ResizablePanels.vue';
 import Vqb from './components/Vqb.vue';
-export { DataViewer, FormRenameStep, Pipeline, ResizablePanels, Vqb };
+export { DataViewer, Pipeline, ResizablePanels, Vqb };
